+++ common/domain/handler_GlobalDomainDisabled_test.go
@@ -72,7 +72,7 @@ func (s *domainHandlerGlobalDomainDisabledSuite) SetupSuite() {
 		log.SetOutput(os.Stdout)
 	}
 
-	s.TestBase = external.NewTestBaseWithExternalCassandra(&persistencetests.TestBaseOptions{
+	s.TestBase = external.NewTestBaseWithPublicCassandra(&persistencetests.TestBaseOptions{
 		ClusterMetadata: cluster.GetTestClusterMetadata(false, false),
 	})
 	s.TestBase.Setup()
