+++ common/persistence/persistence-tests/external/cassandra.go
@@ -26,8 +26,8 @@ import (
 	persistencetests "github.com/uber/cadence/common/persistence/persistence-tests"
 )
 
-// NewTestBaseWithExternalCassandra returns a persistence test base backed by cassandra datastore
+// NewTestBaseWithPublicCassandra returns a persistence test base backed by cassandra datastore
 // It is only being used by testing against external/public Cassandra, which require to load the default gocql client
-func NewTestBaseWithExternalCassandra(options *persistencetests.TestBaseOptions) persistencetests.TestBase {
+func NewTestBaseWithPublicCassandra(options *persistencetests.TestBaseOptions) persistencetests.TestBase {
 	return persistencetests.NewTestBaseWithCassandra(options)
 }
